requirements.txt
brother-ql>=0.9.4
elabapy>=0.6.1
pyyaml>=5.1
Pillow>=7.2.0
tabulate>=0.8.7
qrcode>=6.1
